requirement.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
distlib==0.3.9
Django==5.1.1
django-allauth==64.2.1
django-cors-headers==4.4.0
django-filter==24.3
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
idna==3.10
jwcrypto==1.5.6
oauthlib==3.2.2
pillow==10.4.0
platformdirs==4.3.6
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.4
virtualenvwrapper-win==1.2.7